feat: outsource service connection
