+++ app/src/main/java/com/example/werwoelfle/MainActivity.java
@@ -1,9 +1,6 @@
 package com.example.werwoelfle;
 
-import android.content.ComponentName;
 import android.content.Intent;
-import android.content.ServiceConnection;
-import android.os.IBinder;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.Log;
@@ -13,7 +10,7 @@ import android.widget.Spinner;
 public class MainActivity extends AppCompatActivity {
 
     private static final String LOG_TAG = "Werewolf";
-    private MyServiceConnection conn = null;
+    private GameStateConnection conn = null;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -55,24 +52,5 @@ public class MainActivity extends AppCompatActivity {
     private void initPlayers() {
         Spinner playerDropdown = (Spinner) findViewById(R.id.playerCount);
         conn.getApi().setPlayers(Integer.parseInt(playerDropdown.getSelectedItem().toString()));
-    }
-
-    public static class MyServiceConnection implements ServiceConnection {
-
-        private GameState.IServiceApi api;
-
-        public GameState.IServiceApi getApi() {
-            return api;
-        }
-
-        @Override
-        public void onServiceConnected(ComponentName name, IBinder service) {
-            this.api = (GameState.IServiceApi) service;
-        }
-
-        @Override
-        public void onServiceDisconnected(ComponentName name) {
-            api = null;
-        }
     }
 }
